src/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.18.1
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0
pillow==10.0.0
opencv-python==4.8.0.76
matplotlib==3.7.2
joblib==1.3.2
gunicorn==21.2.0
werkzeug==2.3.7
pydantic==2.3.0
